Projection: Stop materializing forwarded columns (#2247)
Up to now, whenever a Projection SELECT a, a+1 FROM t emitted a newly generate column (here: a+1), it also materialized all columns that it forwarded (here: a). The reason for that is that we do not allow ValueSegments and ReferenceSegments to be mixed within one table.

With this PR, we instead forward the ReferenceSegments and create new ReferenceSegments for the newly generated columns. These point to a new, temporary table. Because we use the EntireChunkPosList, this virtually causes no overhead.

Also does a minor improvement to segment_iterate which causes a segment that is iterated on using an EntireChunkPosLists to use the sequential iterator, not the random access one.