Add support for clang-11 and gcc-10 (#2248)
To enable brew upgrade, I fixed the issues that came up with clang-11. Nothing big, details can be found in the comments.

This does not yet add gcc-10/clang-11 to any dependencies. It also does not fix any issues that clang-tidy-11 might complain about.

This was tested on OS X 10.15.6 and Ubuntu 20.10 dev.