Fix segment access counters, add to JSON (#2251)
## Fix for attribute vectors

So far, the segment access counters ignore the AttributeVectorIterable. This means that scans were significantly underrepresented. For example, in TPC-H Q2, the p_size and p_type scans were underreported:

I added a test that checks if a scan correctly increases the access counters for each encoding type.

## Fix for AccessType state machine

A PosList with the entries [0, 2, 3] was erroneously considered to be sequential, not monotonic. This came up while developing the previously mentioned test.

## Inclusion in benchmark JSON

To be able to generate the heatmaps printed above, `--metrics` now includes the necessary data in the JSON.

## Printing SQL queries

This is something that I quickly built during the development of this and found too valuable to throw away. It's simply `Print::print("SELECT * FROM foo");`.