Add the JCC-H benchmark (#2249)
Adds support for the JCC-H benchmark. This... erm... gives us more options to improve our statistics components ;)

Query 9 does not finish for SF 1 right now. As it works with smaller SFs, I do not consider it an issue of this PR, but of the query plan. Something to fix later.