Change aggregate types again (#2252)
While trying to teach the hash aggregate new tricks, I found that the size of the AggregateResult struct has a significant impact in some queries. For TPC-H Q18 and SF10, there are 1.5M results. With #2246, I reduced the size of AggregateResult to 40. Still, this results in a 60 MB result vector. This PR reduces the size of the struct to 24 bytes. At the same time, it slightly cleans up the code (at least I hope so).

Instead of having "primary" and "secondary" aggregates, as well as a distinct set, the AggregateResult now holds a single accumulator field that stores whatever is needed for the current aggregate function. No heap storage, no small_vector, no frills. Furthermore, this drops the optional in favor of simply taking the number of non-NULL values seen so far.