Parallelize projection operator (#2269)
This PR parallelizes the projection operator. The projection is only executed in parallel if the chunk size (number of rows) lies over a certain threshold. The threshold still needs to be re-evaluated overtime to find the value which gives the best performance. This doesn't seem to be trivial since it can have a different effect on different benchmarks and it depends on the system load and scheduler. For now, a threshold of 500 should intercept the TPC-C (or in general OLTP) with its small number of rows.

Co-authored-by: Bouncner <martin+github@boissier.de>
Co-authored-by: mrks <markus.dreseler@hpi.de>