TableScan: track number of chunks where all rows match (#2283)
While working with @dey4ss on our research seminar topic (what-if clustering cost models), we noticed that the table scan currently tracks only the number of chunks for which the scan was skipped because no row matches. However, this number is often 0 because those chunks get pruned before the scan is executed.
This PR adds an additional counter to track the number of chunks for which the scan was skipped because all rows match. This does happen, e.g., for BETWEEN scans on l_shipdate.

Those two metrics are not identical because for the latter we still need to write the output.

EDIT: As @Bensk1 pointed out, this PR changes the way sorted scans are tracked: If sorted BETWEEN scans use an early exit (no match) or an all-match-shortcut, the scan now increases the respective no/all-match counter, rather than the sorted scan counter, as the scan is not really executed with binary search. ColumnVsValue scans are not affected because they do not have a similar early-exit code path (unless I missed it).