JoinHash: Improve PosList handling of hash map bigly (#2294)
The PosHashTable is one of the most performance-critical data structures. Until now, it holds many small_vector<RowID, 1>. If a value occurs more than once on the build side, we perform many small allocations. With this PR, we draw the memory for these allocations from a thread-local pool that is backed by a monotonic buffer resource. This reduces the stress on the system allocator.

Secondly, we now rewrite this data structure before we start probing. The nature of having small PosLists of variable sizes is that they are scattered all around the memory. This is not particularly friendly towards the caches. As such, we now merge all of these small_vectors into one giant vector and store the offsets of the sublists.