Add threshold for parallel execution in table scan and join hash (#2287)
This PR adds a threshold for the parallel execution of jobs in the build and probe step of the join hash.

It is implemented in the following way:

    A static JOB_SPAWN_THRESHOLD attribute is added to the JoinHashclass
    The jobs that perform the actual materialize, partition by redix, build and probe are executed directly if the number of elements of the partitions lies under the JOB_SPAWN_THRESHOLD

Like in #2269 the threshold still needs to be re-evaluated over time to find the value which gives the best performance. For now, a threshold of 500 is selected.