Projection: don't forward columns that are evaluated in expressions (#2307)
Assuming a Q1-like projection

   l_returnflag,
   sum(l_extendedprice),
   sum(l_extendedprice*(1-l_discount))

Currently, we will forward columns l_returnflag and l_extendedprice (SUM is processed in the aggregation). Additionally, we need to evaluate l_extendedprice*(1-l_discount). For this evaluation, l_extendedprice is materialized (basically an unencoded value segment) and cached.
This PR checks if a forwardable column is materialized and cached in the expression evaluator. If this is the case, we do no longer forward the column and use the expression evaluator.

The projection runtime is virtually unchanged as forwarding comes for free. But the following operators does not need to resolve the forwarded column via a position list, but can directly process it sequentially.

This PR mainly effects Q1 and Q8 (but super short runtime) of TPC-H. All other runtime changes should be noise.
I worked on this issue as I was working on LZ4 peformance. For other encodings than dictionary, the effect for avoiding multiple accesses to the same column via a position list is larger. For LZ4, this branch improves Q1 performance by 7% (both GCC and clang).