Disable binary caches for CI (#2280)
For the join order benchmark, binary tables are written into the table_path. In CI, this path is shared by different targets that are potentially executed concurrently. This sometimes led to issues with corrupted binary files.

I am also sneaking in some things that are too minor to have their own PR.