AggregateHash: For consecutive group by values, group without hashing (#2317)
In some cases (e.g., TPC-H Q18), we aggregate with consecutive values being used as a group by key. Notably, this is the case when aggregating on the serial primary key of a table without filtering the table before. In these cases, we do not need to perform a full hash-based aggregation, but can use the values as immediate indexes into the list of results. To handle smaller gaps, we include cases up to a certain threshold, but at some point these gaps make the approach less beneficial than a proper hash-based approach.