wait_for_tasks: Execute own tasks first (#2325)
Currently, Worker::wait_for_tasks checks if all tasks are completed. If they are not, the worker pulls tasks from the queue and executes them until all tasks that we wait on are done. This, however, leads to a problem if one of the pulled tasks calls wait_for_tasks itself. Over time, the worker builds a stack of wait_for_tasks calls and the first task has to wait until all tasks that the worker started because it was bored are completed.