Fix benchmark_all.sh script (#2331)
Currently, not all TPC-H runs execute properly as the binary cached data is sorted in a clustered run (using --clustering=Pruning) which fails in the next non-clustered run.

This PR changes the clustered run to use --dont_cache_binary_tables. We could move the run to the very end, but that would probably crash with the next benchmark_all.sh run.