Optimizer: Run BetweenComposition and ChunkPruning rules bottom-up (#2284)
This PR removes the legacy function `_apply_to_plan_inputs_without_subqueries`, which was based on recursion and prone to errors when used. (Fixes #2276)

Due to the removal of the above function from `AbstractRule`, the following optimizer rules needed refactoring.

### SubqueryToJoinRule
The SubqueryToJoinRule did not change much. To traverse the LQP from top to bottom, it now uses the `visit_lqp` function and tracks all visited nodes to prevent any revisitation.

### BetweenCompositionRule & ChunkPruningRule
The BetweenCompositionRule and the ChunkPruningRule received a new traversal approach. They now traverse LQPs in a bottom-up manner using `visit_lqp_upwards`. For the ChunkPruningRule, this means that in some queries, more chunks can be pruned than before.