Use the scheduler for controlled data preparation parallelism (#2403)
With this PR, data preparation steps such as encoding and sorting table data, as well as generating table statistics, are executed using a `NodeQueueScheduler` instance. By default, the number of workers for the parallelized data preparation is equal to `std::thread::hardware_concurrency()`. A user can use the optional new benchmark flag `data_preparation_cores` to set the number of workers used for the data preparation.