Consider NULL values properly during selectivity calculation
Previously the selectivity of an operator scan predicate was calculated on basis of the number of values
in the histogram of the corresponding column. As histograms do not hold NULL values and therefore can have
different total counts of values than the input table this can lead to wrong cardinality estimations.
To calculate the final cardinalithy estimation of scan predicate on the input table, the input table
row count is multiplied with the calculated selectivity. Therefore the selectivity should be calculated
on basis of the row count of the input table, too. This actually already happens in estimations for
other operators (e.g. inner_equi_join and semi_join).

Contributed by the DYOD 2021 group of Marc Rosenau, Theresa Hradilak, Leo Wendt, and Furkan Simsek