Add Casts and Date Intervals (#2411)
This PR,enables Hyrise to resolve SQL's `CAST` and `INTERVAL` statements. Intervals are implemented only for dates.
This is achieved by a new `IntervalExpression` and date utilities. Anyway, this PR does not privode a full support of a Date type. Furthermore, the existing implementation of casts was revisited to match the SQL standard.