Jenkinsfile: Add root repository and submodule repositories as safe directories (#2449)
With a recent git security patch, git checks the owner of the top-level directory and does not perform a git command if the owner is not the current user. This check can be disabled for a repository by adding it as a safe directory, i.e., `git config --global --add safe.directory <dir>`.

This PR adds the directory of the root repository and the directories of the submodule repositories as safe directories.