Enable pushdown of predicate barriers (#2464)
The `PredicatePlacementRule` pushes down predicates until it reaches a barrier node that prevents further pushdown. In this case, pushdown predicates are inserted above the barrier node, and the pushdown traversal restarts at the barrier node.

Our `_push_down_traversal` subroutine takes a given node and processes the inputs. However, in some cases, the given node (a barrier from a previous pushdown) is also a predicate eligible for pushdown. Currently, these barrier predicates (having multiple outputs) cannot be pushed down.

This PR fixes the implementation so that predicate barriers are pushed down as well. As a result, the TPC-DS and Join Order benchmarks profit. In addition, the predicate pushdown code for Union-diamonds gets simplified.