Refactor write benchmarks and add threading
* Refactor: Make Write Benchmarks write random values, Reuse NUMBER_OF_BYTES/ELEMENTS

* Extract FileIOWriteMMAPBenchmarks into its own file.

Extracting the MMAP benchmarks into their own files allows for faster and more specific benchmark runs and makes the code leaner. This is especially relevant as we will introduce extra complexity with threaded benchmarks soon.
To not duplicate the Setup(), TearDown() and sanity_check() code it was extracted into the FileIOWriteBenchmark header. The nice way of only declaring those methods in the header and then implementing them in the corresponding cpp and including header and cpp for other build targets that need the code did not work. The benchmarks which get run by google benchmark in the
file_io_write_micro_benchmark.cpp get defined unbounded and including the cpp in any build target will automatically run them. Therefore to be able to have the implementation of SetUp() and co without having to run unwanted benchmarks, the implementation of said methods was put into the header file. This isn't pretty but as we are sure the header will only be included in our dedicated benchmark files I think it's acceptable.

* Make FileIOWriteBenchmark with system calls threaded.

The benchmarks using write and pwrite can now be passed a list of different thread numbers. To benchmark sequential versions (with a thread count of 1) we still use the sequential implementation to avoid measuring unneccessary thread overhead.

* refactored into single/multithread functions and removed threads.clear() call

* Added generate_random_numbers to util and PR Feedback

* introduced close_file_and_fail function in util and PR feedback

* introduced data access mode enum in mmap write and further refactoring to fail_and_close_file

* Fix asserts testing the wrong way around.

* Fix wrong-way-around asserts in file_io_write_mmap_benchmark

* Extract single threaded mmap write & run format script

* Add first version of multi-threaded mmap write benchmarks

* Refactor to use min function according to PR review

* implemented PR feedback: Asserts, Function renaming

* added 1 as thread parameter to MMAP ATOMIC MAP PRIVATE

Co-authored-by: Sandro Speh <sandrospeh@googlemail.com>
Co-authored-by: Sandro Sp <38314662+SanJSp@users.noreply.github.com>