Integrate Partial Indexes (#2448)
This PR implements a partial hash index (`PartialHashIndex`), i.e., a
single-column index structure that can be created for one or more
segments of the corresponding column without the need to index all
segments. This PR is based on the work started by @bengelhaupt,
@Jasperhino, and @vxrahn in #2386.

This `PartialHashIndex` index differs from the already implemented
indexes because it is created over multiple chunks of a table instead of
being on a per-chunk basis. It can be constructed for a set of chunks
and can later be modified by adding additional chunks or removing
already indexed chunks.
The already existing index implementations also share the property of
having the methods `lower_bound` and `upper_bound` used for range
queries. The `PartialHashIndex` is based on an unsorted hash map and,
therefore, cannot provide such functionality efficiently. Instead, it
provides the functions `equals` and `not_equals`, which allow lookups in
constant time.
This is achieved by using a hash map that maps a given value to a vector
of positions (i.e., `RowID`s) where this value is located. The
`FlatMapIterator` allows iterating over these lists of `RowID`s.

Co-authored-by: Tobias Jordan <tobias.jordan@student.hpi.uni-potsdam.de>
Co-authored-by: Ben-Noah Engelhaupt <code@bengelhaupt.com>
Co-authored-by: Vincent Xeno Rahn <vinni@akv.i24.cc>
Co-authored-by: Jasperhino <blum.jasper@gmail.com>