Improve output writing of hash-based aggregate (#2612)
Improve the multi-threading performance and output structure of the
hash aggregate. Instead of writing a large single chunk as the result, the
group-by columns and aggregation results are now split into chunks
(using `Chunk::DEFAULT_SIZE`). There are some improvements for parallel
query executions. On the one hand, writing is parallelized. On the other
hand, succeeding operators might perform better as they don't execute on
a huge single chunks (most operators parallelize on chunks; for
single-threaded execution, more chunks can decrease performance (see
TPC-DS and comments below)).

Instead of materializing the group-by columns, we only forward
reference columns. This is done the same way the projection operator
works. As we need to combine reference and value segments (i.e.,
group-by columns and aggregate results), we create intermediate value
segments which are referenced in the output table.

Co-authored-by: Daniel Lindner <27929897+dey4ss@users.noreply.github.com>