Encode tables from SQL import (#2679)
Update the SQL-Parser to allow the following syntax:  
`COPY a_table FROM 'a_file.tbl' WITH (FORMAT BIN, ENCODING
'Unencoded');`
This way it is possible to specify the encoding of a table that is
loaded via `COPY`.

How the encoding of a Table is determined:
Binary Files keep their encoding.
If no encoding is specified, the imported table uses the default
encoding of `SegmentEncodingSpec` (Dictionary).
If a column does not support the specified encoding, it is encoded using
`Dictionary` encoding.

---------

Co-authored-by: Daniel Lindner <27929897+dey4ss@users.noreply.github.com>