[DYOD] Maintain UCCs in face of potentially changing data (#2599)
Use only UCCs in the LQP that are known to be valid. Similarly, the UCCDiscoveryPlugin also only revalidates UCCs that are neither known to be valid nor invalid. Note that plans that were optimized using a UCC that is not schema-given are as of now not distinguishable. This will be handled in #2600 and #2698. 

---------

Co-authored-by: kalathen <kalathen@jyu.fi>
Co-authored-by: Margarete <margarete.dippel@student.hpi.de>
Co-authored-by: Robert Weeke <Robert.Weeke@student.hpi.uni-potsdam.de>
Co-authored-by: Robert <93660664+Rob2U@users.noreply.github.com>
Co-authored-by: Daniel Lindner <27929897+dey4ss@users.noreply.github.com>