Allow copy from table without JSON meta file (#2695)
This branch implements COPY FROM a .csv file into an existing table
without requiring a .csv.json file. The csv file has to be separated by
comma for this to work.

The csv parser now no longer searches for a .csv.json file itself.
Instead, the caller to CsvReader::parse has to load the json file
themselves. Since we want to move away from json files, it makes
sense for the csv parser to not search for a json file automatically anymore.